+++ ge/org.osate.ge/src/org/osate/ge/internal/ui/handlers/AlignLeftHandler.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
+import org.osate.ge.graphics.Point;
 import org.osate.ge.internal.diagram.runtime.AgeDiagram;
 import org.osate.ge.internal.diagram.runtime.DiagramElement;
-import org.osate.ge.internal.diagram.runtime.Point;
 import org.osate.ge.internal.ui.util.UiUtil;
 
 public class AlignLeftHandler extends AbstractHandler {
@@ -23,7 +23,7 @@ public class AlignLeftHandler extends AbstractHandler {
 		}
 
 		diagram.modify("Align Left", m -> {
-			int left = AgeHandlerUtil.getPrimaryDiagramElement(selectedDiagramElements).getX();
+			double left = AgeHandlerUtil.getPrimaryDiagramElement(selectedDiagramElements).getX();
 			for (final DiagramElement de : selectedDiagramElements) {
 				m.setPosition(de, new Point(left, de.getY()));
 			}
